Consolidate the two separate clip making views into one
